Add script for consistent tagging
My workflow when updating the kernel is to rebase the x230 branch
against master, resolve the conflicts and then rebuild the package.

As I'm rebasing the branch, the commits from which I've successfully
built kernels would be lost if I didn't tag them.

This script generates git tags in the same format that the upstream repo
uses, and adds a simple one line message to explain what the tag is for.