Remove virtualbox guest support from linux-current
This removes the VirtualBox guest modules support which are currently just
not complete or ready for a full replacement of the standalone package.
Additionally we remove the DRM driver for VirtualBox as its long causing
us issues with fuzzy/broken X on boot in guest-addition-less ISOs, so
now we'll restore parity with the `linux-lts` kernel.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>